src/requirements.txt
cached-property==1.5.1
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
dnspython==1.16.0
exchangelib==3.2.0
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
isodate==0.6.0
lxml==4.5.1
ntlm-auth==1.4.0
oauthlib==3.1.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
